Add EPPlus package, refine CSS, and update views
Added the EPPlus package to the HealthMonitorApp for handling Excel files. CSS styles have been adjusted for a more consistent layout. Expanded functionality in Create view to include excluded controllers and methods. In the Details view, an option was added to download summary as Excel. In the _DeleteConfirmation view, detailed information about the associated API groups and endpoints to be deleted was included.